⚡️ refactor: Blogs Sections improvement
