+++ src/pages/index.js
@@ -3,8 +3,8 @@ export { default as PolicyPage } from "./policity/PolicyPage";
 export { default as LoginPage } from "./LoginPage";
 // export { default as MainLayout } from "./layout/MainLayout"; // Movido a shared
 // export { default as AdminLayout } from "./layout/AdminLayout"; // Movido a shared
-export { default as BlogPage } from "./blog/BlogPage";
-export { default as BlogDetailsPage } from "./blog/BlogDetailsPage";
+// export { default as BlogPage } from "./blog/BlogPage"; // Moved to features/blogs
+// export { default as BlogDetailsPage } from "./blog/BlogDetailsPage"; // Moved to features/blogs
 export { default as BlogAdminPage } from "./blogAdmin/BlogAdminPage";
 export { default as PolicyAdminPage } from "./policyAdmin/PolicyAdminPage";
 export { default as ServiceAdminPage } from "./serviceAdmin/ServiceAdminPage";
